Restructured solution and projects
