--- MonkeyInterpreter.AST/IExpression.cs
@@ -1,4 +1,0 @@
-﻿namespace MonkeyInterpreter.AST;
-public interface IExpression : INode
-{
-}
