--- MonkeyInterpreter.AST/IStatement.cs
@@ -1,6 +1,0 @@
-﻿namespace MonkeyInterpreter.AST;
-
-public interface IStatement : INode
-{
-	
-}
