+++ MonkeyInterpreter.Core/AbstractSyntaxTree/BooleanLiteral.cs
@@ -1,6 +1,6 @@
-﻿using MonkeyInterpreter.Core;
+﻿using MonkeyInterpreter.Core.Parser;
 
-namespace MonkeyInterpreter.AST;
+namespace MonkeyInterpreter.Core.AbstractSyntaxTree;
 
 public class BooleanLiteral : IExpression
 {
