+++ MonkeyInterpreter.Core/AbstractSyntaxTree/ExpressionStatement.cs
@@ -1,6 +1,6 @@
-﻿using MonkeyInterpreter.Core;
+﻿using MonkeyInterpreter.Core.Parser;
 
-namespace MonkeyInterpreter.AST;
+namespace MonkeyInterpreter.Core.AbstractSyntaxTree;
 
 public class ExpressionStatement : IStatement
 {
