+++ MonkeyInterpreter.Core/AbstractSyntaxTree/IExpression.cs
@@ -1,0 +1,4 @@
+﻿namespace MonkeyInterpreter.Core.AbstractSyntaxTree;
+public interface IExpression : INode
+{
+}
