+++ MonkeyInterpreter.Core/AbstractSyntaxTree/IStatement.cs
@@ -1,0 +1,6 @@
+﻿namespace MonkeyInterpreter.Core.AbstractSyntaxTree;
+
+public interface IStatement : INode
+{
+	
+}
