+++ MonkeyInterpreter.Core/Evaluator/BooleanObject.cs
@@ -1,4 +1,4 @@
-﻿namespace MonkeyInterpreter.Core;
+﻿namespace MonkeyInterpreter.Core.Evaluator;
 
 public class BooleanObject : IObject
 {
