+++ MonkeyInterpreter.Core/Evaluator/IntegerObject.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace MonkeyInterpreter.Core;
+namespace MonkeyInterpreter.Core.Evaluator;
 
 public class IntegerObject : IObject
 {
