+++ MonkeyInterpreter.REPL/MonkeyInterpreter.REPL.csproj
@@ -4,10 +4,10 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <OutputType>Exe</OutputType>
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\MonkeyInterpreter.AST\MonkeyInterpreter.AST.csproj" />
       <ProjectReference Include="..\MonkeyInterpreter.Core\MonkeyInterpreter.Core.csproj" />
     </ItemGroup>
