+++ MonkeyInterpreter.Tests/Parser/AstTest.cs
@@ -1,5 +1,6 @@
-﻿using MonkeyInterpreter.AST;
-using MonkeyInterpreter.Core;
+﻿using MonkeyInterpreter.Core;
+using MonkeyInterpreter.Core.AbstractSyntaxTree;
+using MonkeyInterpreter.Core.Parser;
 
 namespace MonkeyInterpreter.Tests.Parser;
